Small refactoring steps - start extracting polkadot integration, remove bot/server distinct specs
