+++ .gitignore
@@ -6,5 +6,4 @@ sqlite.db
 build
 
 # Autogenerated
-env.bot.config.json.d.ts
-env.server.config.json.d.ts
+env.*.config.json.d.ts
