--- env.bot.config.json
@@ -1,58 +1,0 @@
-{
-  "SMF": {
-    "BOT": {
-      "BACKEND_URL": {
-        "description": "full url for the bot to reach the backend",
-        "default": "http://localhost:5555",
-        "type": "string"
-      },
-      "DRIP_AMOUNT": {
-        "description": "default amount of token to send",
-        "default": 0.5,
-        "type": "number"
-      },
-      "MATRIX_ACCESS_TOKEN": {
-        "description": "your bot access token here is how to find it https://t2bot.io/docs/access_tokens/.",
-        "required": true,
-        "masked": true,
-        "type": "string"
-      },
-      "MATRIX_BOT_USER_ID": {
-        "description": "your bot user id",
-        "required": true,
-        "regexp": "^@.*:.*$",
-        "type": "string"
-      },
-      "MATRIX_SERVER": {
-        "description": "matrix server url",
-        "default": "https://matrix.org",
-        "type": "string"
-      },
-      "NETWORK_DECIMALS": {
-        "description": "decimal amount for the network",
-        "default": 12,
-        "type": "number"
-      },
-      "NETWORK_UNIT": {
-        "description": "token unit for the network",
-        "default": "UNIT",
-        "type": "string"
-      },
-      "FAUCET_IGNORE_LIST": {
-        "description": "A list of Matrix accounts that will be silently (but logged) ignored, comma separated. Example: \"@alice:matrix.org,@bob:domain.com\"",
-        "default": "",
-        "type": "string"
-      },
-      "DEPLOYED_REF": {
-        "description": "git ref which of deployed bot",
-        "default": "unset",
-        "type": "string"
-      },
-      "DEPLOYED_TIME": {
-        "description": "time when we deployed bot",
-        "default": "unset",
-        "type": "string"
-      }
-    }
-  }
-}
