--- env.server.config.json
@@ -1,63 +1,0 @@
-{
-  "SMF": {
-    "BACKEND": {
-      "FAUCET_ACCOUNT_MNEMONIC": {
-        "description": "mnemonic seed from faucet account (create via polkadot.js.org)",
-        "mandatory": true,
-        "masked": true,
-        "type": "string"
-      },
-      "FAUCET_BALANCE_CAP": {
-        "description": "Upper limit cap on whether or not the account can receive more tokens. Defaults to 100.",
-        "default": 100,
-        "type": "number"
-      },
-      "INJECTED_TYPES": {
-        "description": "if any type must be overridden",
-        "default": {},
-        "type": "string"
-      },
-      "NETWORK_DECIMALS": {
-        "description": "decimal amount for the network",
-        "default": 12,
-        "type": "number"
-      },
-      "PORT": {
-        "description": "the port you want the server to listen on (should be part of the bot's `SMF_BOT_BACKEND_URL`.)",
-        "default": 5555,
-        "type": "number"
-      },
-      "RPC_ENDPOINT": {
-        "description": "ws rpc node endpoint",
-        "default": "https://westend-rpc.polkadot.io/",
-        "type": "string"
-      },
-      "DEPLOYED_REF": {
-        "description": "git ref which of deployed app",
-        "default": "unset",
-        "type": "string"
-      },
-      "DEPLOYED_TIME": {
-        "description": "time when we deployed app backend",
-        "default": "unset",
-        "type": "string"
-      },
-      "EXTERNAL_ACCESS": {
-        "description": "Whether the backend should serve external drip requests",
-        "default": "false",
-        "type": "boolean"
-      },
-      "DRIP_AMOUNT": {
-        "description": "default amount of token to send",
-        "default": "0.5",
-        "type": "string"
-      },
-      "RECAPTCHA_SECRET": {
-        "description": "A secret recaptcha token used to validate external requests",
-        "default": "",
-        "masked": true,
-        "type": "string"
-      }
-    }
-  }
-}
