+++ src/dripper/polkadotApi.ts
@@ -2,10 +2,10 @@ import "@polkadot/api-augment";
 import { ApiPromise } from "@polkadot/api";
 import { HttpProvider } from "@polkadot/rpc-provider";
 
-import { config } from "./config";
+import { serverConfig } from "../config";
 
-const rpcEndpointUrl = config.Get("RPC_ENDPOINT");
-const injectedTypes = JSON.parse(config.Get("INJECTED_TYPES")) as Record<string, string>;
+const rpcEndpointUrl = serverConfig.Get("RPC_ENDPOINT");
+const injectedTypes = JSON.parse(serverConfig.Get("INJECTED_TYPES")) as Record<string, string>;
 
 const provider = new HttpProvider(rpcEndpointUrl);
 const types = injectedTypes;
